refactor: migrate Ethereum wallets to shared base class and remove MetaMask SDK
- Create EthereumWalletBase abstract class using EIP-6963/mipd for wallet detection
- Migrate MetaMask, Rabby, and Base wallets to extend shared base class
- Remove @metamask/sdk dependency to reduce bundle size by several MB
- Eliminate ~970 lines of duplicated code across wallet implementations
- Standardize all Ethereum wallets to use the same low-level provider API

This change significantly reduces bundle size by removing the heavyweight MetaMask SDK
and its dependencies (Socket.io, QR code libraries, ECIES encryption) while maintaining
full functionality through direct EIP-6963 provider access.
