+++ packages/controller/src/wallets/index.ts
@@ -1,6 +1,8 @@
 export * from "./argent";
+export * from "./base";
 export * from "./braavos";
 export * from "./bridge";
+export * from "./ethereum-base";
 export * from "./metamask";
 export * from "./phantom";
 export * from "./rabby";
